Module alpha_vantage::api

Module for basic definition of user information like setting API and requesting through that API

Structs

ApiClient

Struct for initializing client which contains different method for API call

Enums

OutputSize

Enum for declaring output size of API call

Provider

Provider for alpha vantage API

TimeSeriesInterval

Enum for declaring interval for intraday time series