Struct alpha_vantage::api::ApiClient

pub struct ApiClient<'a> { /* fields omitted */ }

Struct for initializing client which contains different method for API call

Implementations

impl<'a> ApiClient<'a>

pub fn set_api<T>(api: &'a str, client: T) -> Self where
T: HttpClient + 'static,

Method for initializing ApiClient struct using user provided client

use alpha_vantage::api::ApiClient;
let api = ApiClient::set_api("some_key", reqwest::Client::new());

pub fn get_api_key(&self) -> &str

Method to get api key

use alpha_vantage::api::ApiClient;
let api = alpha_vantage::api::ApiClient::set_api("some_key", reqwest::Client::new());
assert_eq!(api.get_api_key(), "some_key");

pub fn crypto(
&'a self,
function: CryptoFunction,
symbol: &'a str,
market: &'a str
) -> CryptoBuilder<'a>

Crypto method for calling cryptography function with help of CryptoBuilder

Example

#[tokio::main]
async fn main() {
    let api = alpha_vantage::set_api("demo", reqwest::Client::new());
    let crypto = api
        .crypto(alpha_vantage::utils::CryptoFunction::Daily, "BTC", "CNY")
        .json()
        .await
        .unwrap();
    assert_eq!(crypto.digital_code(), "BTC");
    assert_eq!(crypto.digital_name(), "Bitcoin");
    assert_eq!(crypto.market_code(), "CNY");
    assert_eq!(crypto.market_name(), "Chinese Yuan");
}

pub fn custom(&'a self, function: &'a str) -> CustomBuilder<'a>

Method for calling custom function not implemented currently in library using CustomBuilder

pub fn earning(&'a self, symbol: &'a str) -> EarningBuilder<'a>

Method for returning EarningBuilder for earning API

Example

#[tokio::main]
async fn main() {
    let api = alpha_vantage::set_api("demo", reqwest::Client::new());
    let earning = api.earning("IBM").json().await.unwrap();
    let symbol = earning.symbol();
    assert_eq!(symbol, "IBM");
}

pub fn exchange(
&'a self,
from_currency: &'a str,
to_currency: &'a str
) -> ExchangeBuilder<'a>

Method for creating ExchangeBuilder for exchanging currency value from one currency to another currency.

Example

#[tokio::main]
async fn main() {
    let api = alpha_vantage::set_api("demo", reqwest::Client::new());
    let exchange = api.exchange("BTC", "CNY").json().await.unwrap();
    assert_eq!(exchange.name_from(), "Bitcoin");
    assert_eq!(exchange.code_from(), "BTC");
    assert_eq!(exchange.name_to(), "Chinese Yuan");
    assert_eq!(exchange.code_to(), "CNY");
}

pub fn forex(
&'a self,
function: ForexFunction,
from_symbol: &'a str,
to_symbol: &'a str,
interval: TimeSeriesInterval,
output_size: OutputSize
) -> ForexBuilder<'a>

Method for creating ForexBuilder for Forex API

Example

use alpha_vantage::utils::*;
#[tokio::main]
async fn main() {
    let api = alpha_vantage::set_api("demo", reqwest::Client::new());
    let forex = api
        .forex(
            ForexFunction::Weekly,
            "EUR",
            "USD",
            TimeSeriesInterval::None,
            OutputSize::None,
        )
        .json()
        .await
        .unwrap();
    assert_eq!(forex.symbol_from(), "EUR");
    assert_eq!(forex.symbol_to(), "USD");
    assert!(forex.interval().is_none());
}

pub fn quote(&'a self, symbol: &'a str) -> QuoteBuilder<'a>

Method for creating QuoteBuilder from APIClient

Example

#[tokio::main]
async fn main() {
    let api = alpha_vantage::set_api("demo", reqwest::Client::new());
    let quote = api.quote("MSFT").json().await.unwrap();
    let symbol = quote.symbol();
    assert_eq!(symbol, "MSFT");
}

pub fn search(&'a self, keywords: &'a str) -> SearchBuilder<'a>

Method for creating search builder

Example

#[tokio::main]
async fn main() {
    let api = alpha_vantage::set_api("demo", reqwest::Client::new());
    let search = api.search("BA").json().await.unwrap();
    let first_search_result = &search.result()[0];
    assert_eq!(first_search_result.symbol(), "BA");
    assert_eq!(first_search_result.name(), "Boeing Company");
    assert_eq!(first_search_result.stock_type(), "Equity");
    assert_eq!(first_search_result.region(), "United States");
    assert_eq!(first_search_result.currency(), "USD");
    assert_eq!(first_search_result.match_score(), 1.0);
}

pub fn sector(&'a self) -> SectorBuilder<'a>

Method for creating SectorBuilder

Example

#[tokio::main]
async fn main() {
    let api = alpha_vantage::set_api("demo", reqwest::Client::new());
    let sector = api.sector().json().await.unwrap();
    assert_eq!(
        sector.information(),
        "US Sector Performance (realtime & historical)"
    );
}

pub fn stock_time(
&'a self,
function: StockFunction,
symbol: &'a str,
interval: TimeSeriesInterval,
output_size: OutputSize
) -> TimeSeriesBuilder<'a>

Method for creating Stock time Builder from APIClient

Example

use alpha_vantage::utils::*;
#[tokio::main]
async fn main() {
    let api = alpha_vantage::set_api("demo", reqwest::Client::new());
    let stock = api
        .stock_time(
            StockFunction::Weekly,
            "MSFT",
            TimeSeriesInterval::None,
            OutputSize::None,
        )
        .json()
        .await
        .unwrap();
    assert_eq!(stock.symbol(), "MSFT");
    assert!(stock.interval().is_none());
}

pub fn technical_indicator(
&'a self,
function: &'a str,
symbol: &'a str,
interval: TechnicalIndicatorInterval
) -> IndicatorBuilder<'a>

Method for technical indicator builder

Example

#[tokio::main]
async fn main() {
    let api = alpha_vantage::set_api("demo", reqwest::Client::new());
    let technical = api
        .technical_indicator(
            "SMA",
            "IBM",
            alpha_vantage::utils::TechnicalIndicatorInterval::Weekly,
        )
        .time_period(10)
        .series_type("open")
        .json()
        .await;
    assert!(technical.is_ok());
}