Trait alpha_vantage::client::HttpClient

pub trait HttpClient {
    fn get_alpha_vantage_provider_output<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        path: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<String>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    fn get_rapid_api_provider_output<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self, 
        path: &'life1 str, 
        api_key: &'life2 str
    ) -> Pin<Box<dyn Future<Output = Result<String>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;
}

Trait which can be implemented for all common library client for getting output from server surf and reqwest are two client which are supported with feature flag. If you prefer alternate http client you can add support by implementing HttpClient trait for client. Some example of other client which can be used are isahc client

Required methods

fn get_alpha_vantage_provider_output<'life0, 'life1, 'async_trait>(
&'life0 self,
path: &'life1 str
) -> Pin<Box<dyn Future<Output = Result<String>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

AlphaVantage provider output function which provides one field path where get GET request needs to be performed

fn get_rapid_api_provider_output<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
path: &'life1 str,
api_key: &'life2 str
) -> Pin<Box<dyn Future<Output = Result<String>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

RapidAPI provider function which provides two field path and api_key. Path needs to be set along with header x-rapidapi-host as alpha-vantage.p.rapidapi.com and header x-rapidapi-key same as api_key field