Module alpha_vantage::stock_time

Module for stock time series

This suite of APIs provide realtime and historical global equity data in 4 different temporal resolutions: (1) daily, (2) weekly, (3) monthly, and (4) intraday. Daily, weekly, and monthly time series contain 20+ years of historical data

You can read about Stock Time API and what it returns on alphavantage documentation

Structs

Data

Struct for Data value

MetaData

Struct for storing Meta Data value

TimeSeries

Struct for storing time series data

TimeSeriesBuilder

Builder to create new TimeSeries

Enums

StockFunction

Enum for declaring function for stock time series by defining which type of series of stock to be returned