Trait alpha_vantage::stock_time::VecEntry

pub trait VecEntry {
    fn find(&self, time: &str) -> Option<Entry>;

    fn latest(&self) -> Entry;

    fn latestn(&self, n: usize) -> Result<Vec<Entry>>;
}

trait which helps for performing some common operation on Vec

Required methods

fn find(&self, time: &str) -> Option<Entry>

Find a entry with a given time as a input return none if no entry found

fn latest(&self) -> Entry

Return a entry which is of latest time period

fn latestn(&self, n: usize) -> Result<Vec<Entry>>

Return a top n latest Entry if n Entry is present else return Error