Module alpha_vantage::technical_indicator

Module for Technical Indicator

Technical indicator values are updated realtime: the latest data point is derived from the current trading day of a given equity or currency exchange pair.

You can read about Technical Indicator API and what it returns on alphavantage documentation

Structs

DataCollector
Struct for storing a data values

Indicator
Struct for indicator

Functions

technical_indicator
Function used to create a [Indicator][Indicator] struct.