Struct alpha_vantage::user::APIKey

pub struct APIKey { /* fields omitted */ }

Struct for initializing api key value as well as contain different method for API call

Methods

impl APIKey

#[must_use]pub fn set_api(api: &str) -> Self

Method for initializing [APIKey][APIKey] struct

use alpha_vantage::user::APIKey;
let api = APIKey::set_api("some_key");

#[must_use]pub fn set_with_timeout(api: &str, timeout: u64) -> Self

Set API value with timeout period

use alpha_vantage::user::APIKey;
let api_with_custom_timeout = APIKey::set_with_timeout("your_api_key", 45);

#[must_use]pub fn set_with_env(env_name: &str) -> Self

Set out [APIKey][APIKey] by reading out environment variable

use alpha_vantage::user::APIKey;
std::env::set_var("KEY_NAME", "some_key");
let api_from_env = APIKey::set_with_env("KEY_NAME");
assert_eq!(api_from_env.get_api(), "some_key");

pub fn update_timeout(&mut self, timeout: u64)

Update timeout for API key

use alpha_vantage::user::APIKey;
let mut api = alpha_vantage::user::APIKey::set_api("some_key");
assert_eq!(api.get_timeout(), 30_u64);
api.update_timeout(60_u64);
assert_eq!(api.get_timeout(), 60_u64);

#[must_use]pub fn get_api(&self) -> &str

Method to get api key

use alpha_vantage::user::APIKey;
let api = alpha_vantage::user::APIKey::set_api("some_key");
assert_eq!(api.get_api(), "some_key");

#[must_use]pub fn get_timeout(&self) -> u64

Get API timeout period

use alpha_vantage::user::APIKey;
let api_with_custom_timeout = APIKey::set_with_timeout("your_api_key", 45);
assert_eq!(api_with_custom_timeout.get_timeout(), 45_u64);

pub async fn crypto_rating<'_, '_>(
&'_ self,
symbol: &'_ str
) -> Result<CryptoRating>

Method for getting crypto health rating

Example

use tokio::prelude::*;
#[tokio::main]
async fn main() {
    let api = alpha_vantage::set_api("demo");
    assert_eq!(api.crypto_rating("BTC").await.unwrap().name(), "Bitcoin");
}

pub async fn crypto<'_, '_, '_>(
&'_ self,
function: CryptoFunction,
symbol: &'_ str,
market: &'_ str
) -> Result<Crypto>

Crypto method for calling cryptography function

Example

use tokio::prelude::*;
#[tokio::main]
async fn main() {
    let api = alpha_vantage::set_api("demo");
    let crypto = api
        .crypto(alpha_vantage::util::CryptoFunction::Daily, "BTC", "CNY")
        .await
        .unwrap();
    let digital_name = crypto.digital_name();
    assert_eq!(digital_name, "Bitcoin");
}

pub async fn exchange<'_, '_, '_>(
&'_ self,
from_currency: &'_ str,
to_currency: &'_ str
) -> Result<Exchange>

Method for exchanging currency value from one currency to another currency.

Example

use tokio::prelude::*;
#[tokio::main]
async fn main() {
    let api = alpha_vantage::set_api("demo");
    assert_eq!(
        api.exchange("BTC", "CNY").await.unwrap().name_from(),
        "Bitcoin"
    );
}

pub async fn forex<'_, '_, '_>(
&'_ self,
function: ForexFunction,
from_symbol: &'_ str,
to_symbol: &'_ str,
interval: TimeSeriesInterval,
output_size: OutputSize
) -> Result<Forex>

Forex method for calling stock time series

Example

use alpha_vantage::util::*;
use tokio::prelude::*;
#[tokio::main]
async fn main() {
    let api = alpha_vantage::set_api("demo");
    let forex = api
        .forex(
            ForexFunction::Weekly,
            "EUR",
            "USD",
            TimeSeriesInterval::None,
            OutputSize::None,
        )
        .await;
    assert_eq!(forex.unwrap().symbol_from(), "EUR");
}

pub async fn quote<'_, '_>(&'_ self, symbol: &'_ str) -> Result<Quote>

Method for returning Quote Struct

Example

use tokio::prelude::*;
#[tokio::main]
async fn main() {
    let api = alpha_vantage::set_api("demo");
    let quote = api.quote("MSFT").await.unwrap();
    let symbol = quote.symbol();
    assert_eq!(symbol, "MSFT");
}

pub async fn search<'_, '_>(&'_ self, keywords: &'_ str) -> Result<Search>

Search method for searching keyword or company

Example

use tokio::prelude::*;
#[tokio::main]
async fn main() {
    let api = alpha_vantage::set_api("demo");
    let search = api.search("BA").await.unwrap();
    assert_eq!(search.result()[0].symbol(), "BA");
}

pub async fn sector<'_>(&'_ self) -> Result<Sector>

Method for returning out a sector data as struct

Example

use tokio::prelude::*;
#[tokio::main]
async fn main() {
    let api = alpha_vantage::set_api("demo");
    let sector = api.sector().await.unwrap();
    assert_eq!(
        sector.information(),
        "US Sector Performance (realtime & historical)"
    );
}

pub async fn stock_time<'_, '_>(
&'_ self,
function: StockFunction,
symbol: &'_ str,
interval: TimeSeriesInterval,
output_size: OutputSize
) -> Result<TimeSeries>

Stock time method for calling stock time series API

Example

use alpha_vantage::util::*;
use tokio::prelude::*;
#[tokio::main]
async fn main() {
    let api = alpha_vantage::set_api("demo");
    let stock = api
        .stock_time(
            StockFunction::Weekly,
            "MSFT",
            TimeSeriesInterval::None,
            OutputSize::None,
        )
        .await
        .unwrap();
    assert_eq!(stock.symbol(), "MSFT".to_string());
}

pub async fn technical_indicator<'_, '_, '_, '_>(
&'_ self,
function: &'_ str,
symbol: &'_ str,
interval: TechnicalIndicatorInterval,
time_period: Option<u64>,
series_type: Option<&'_ str>,
temporary_value: Vec<TechnicalIndicator>
) -> Result<Indicator>

Technical indicator API caller method

Example

use tokio::prelude::*;
#[tokio::main]
async fn main() {
    let api = alpha_vantage::set_api("demo");
    let technical = api
        .technical_indicator(
            "SMA",
            "IBM",
            alpha_vantage::util::TechnicalIndicatorInterval::Weekly,
            Some(10),
            Some("open"),
            vec![],
        )
        .await;
    assert_eq!(technical.is_ok(), true);
}