Enum alpha_vantage::utils::TechnicalIndicator

pub enum TechnicalIndicator {

    Acceleration(f32),
    Fastdmatype(u32),
    Fastdperiod(u32),
    Fastkperiod(u32),
    Fastlimit(f32),
    Fastmatype(u32),
    Fastperiod(u32),
    Matype(u32),
    Maximum(f32),
    Nbdevdn(u32),
    Nbdevup(u32),
    Signalmatype(u32),
    Signalperiod(u32),
    Slowdmatype(u32),
    Slowdperiod(u32),
    Slowkmatype(u32),
    Slowkperiod(u32),
    Slowlimit(f32),
    Slowmatype(u32),
    Slowperiod(u32),
    Timeperiod1(u32),
    Timeperiod2(u32),
    Timeperiod3(u32),

}

Enum for declaring different optional value of Technical Indicator used so user can pass different optional value as same datatype

Variants

Acceleration(f32)

The acceleration factor. Positive floats are accepted. By default, acceleration=0.01.

Fastdmatype(u32)

Moving average type for the fastd moving average. By default, fastdmatype=0. Integers 0 - 8 are accepted with the following mappings. 0 = Simple Moving Average (SMA), 1 = Exponential Moving Average (EMA), 2 = Weighted Moving Average (WMA), 3 = Double Exponential Moving Average (DEMA), 4 = Triple Exponential Moving Average (TEMA), 5 = Triangular Moving Average (TRIMA), 6 = T3 Moving Average, 7 = Kaufman Adaptive Moving Average (KAMA), 8 = MESA Adaptive Moving Average (MAMA).

Fastdperiod(u32)

The time period of the fastd moving average. Positive integers are accepted. By default, fastdperiod=3

Fastkperiod(u32)

The time period of the fastk moving average. Positive integers are accepted. By default, fastkperiod=5

Fastlimit(f32)

Positive floats are accepted. By default, fastlimit=0.01.

Fastmatype(u32)

Moving average type for the faster moving average. By default, fastmatype=0. Integers 0 - 8 are accepted with the following mappings. 0 = Simple Moving Average (SMA), 1 = Exponential Moving Average (EMA), 2 = Weighted Moving Average (WMA), 3 = Double Exponential Moving Average (DEMA), 4 = Triple Exponential Moving Average (TEMA), 5 = Triangular Moving Average (TRIMA), 6 = T3 Moving Average, 7 = Kaufman Adaptive Moving Average (KAMA), 8 = MESA Adaptive Moving Average (MAMA).

Fastperiod(u32)

Positive integers are accepted. By default, fastperiod=12.

Matype(u32)

Moving average type. By default, matype=0. Integers 0 - 8 are accepted with the following mappings. 0 = Simple Moving Average (SMA), 1 = Exponential Moving Average (EMA), 2 = Weighted Moving Average (WMA), 3 = Double Exponential Moving Average (DEMA), 4 = Triple Exponential Moving Average (TEMA), 5 = Triangular Moving Average (TRIMA), 6 = T3 Moving Average, 7 = Kaufman Adaptive Moving Average (KAMA), 8 = MESA Adaptive Moving Average (MAMA).

Maximum(f32)

The acceleration factor maximum value. Positive floats are accepted. By default, maximum=0.20

Nbdevdn(u32)

The standard deviation multiplier of the lower band. Positive integers are accepted. By default, nbdevdn=2.

Nbdevup(u32)

The standard deviation multiplier of the upper band. Positive integers are accepted. By default, nbdevup=2

Signalmatype(u32)

Moving average type for the signal moving average. By default, signalmatype=0. Integers 0 - 8 are accepted with the following mappings. 0 = Simple Moving Average (SMA), 1 = Exponential Moving Average (EMA), 2 = Weighted Moving Average (WMA), 3 = Double Exponential Moving Average (DEMA), 4 = Triple Exponential Moving Average (TEMA), 5 = Triangular Moving Average (TRIMA), 6 = T3 Moving Average, 7 = Kaufman Adaptive Moving Average (KAMA), 8 = MESA Adaptive Moving Average (MAMA).

Signalperiod(u32)

Positive integers are accepted. By default, signalperiod=9.

Slowdmatype(u32)

Moving average type for the slowd moving average. By default, slowdmatype=0. Integers 0 - 8 are accepted with the following mappings. 0 = Simple Moving Average (SMA), 1 = Exponential Moving Average (EMA), 2 = Weighted Moving Average (WMA), 3 = Double Exponential Moving Average (DEMA), 4 = Triple Exponential Moving Average (TEMA), 5 = Triangular Moving Average (TRIMA), 6 = T3 Moving Average, 7 = Kaufman Adaptive Moving Average (KAMA), 8 = MESA Adaptive Moving Average (MAMA).

Slowdperiod(u32)

The time period of the slowd moving average. Positive integers are accepted. By default, slowdperiod=3.

Slowkmatype(u32)

Moving average type for the slowk moving average. By default, slowkmatype=0. Integers 0 - 8 are accepted with the following mappings. 0 = Simple Moving Average (SMA), 1 = Exponential Moving Average (EMA), 2 = Weighted Moving Average (WMA), 3 = Double Exponential Moving Average (DEMA), 4 = Triple Exponential Moving Average (TEMA), 5 = Triangular Moving Average (TRIMA), 6 = T3 Moving Average, 7 = Kaufman Adaptive Moving Average (KAMA), 8 = MESA Adaptive Moving Average (MAMA).

Slowkperiod(u32)

The time period of the slowk moving average. Positive integers are accepted. By default, slowkperiod=3.

Slowlimit(f32)

Positive floats are accepted. By default, slowlimit=0.01.

Slowmatype(u32)

Moving average type for the slower moving average. By default, slowmatype=0. Integers 0 - 8 are accepted with the following mappings. 0 = Simple Moving Average (SMA), 1 = Exponential Moving Average (EMA), 2 = Weighted Moving Average (WMA), 3 = Double Exponential Moving Average (DEMA), 4 = Triple Exponential Moving Average (TEMA), 5 = Triangular Moving Average (TRIMA), 6 = T3 Moving Average, 7 = Kaufman Adaptive Moving Average (KAMA), 8 = MESA Adaptive Moving Average (MAMA).

Slowperiod(u32)

Positive integers are accepted. By default, slowperiod=26.

Timeperiod1(u32)

The first time period for the indicator. Positive integers are accepted. By default, timeperiod1=7.

Timeperiod2(u32)

The second time period for the indicator. Positive integers are accepted. By default, timeperiod2=14.

Timeperiod3(u32)

The third time period for the indicator. Positive integers are accepted. By default, timeperiod3=28.