Crate alpha_vantage

Rust Client/Wrapper built for Alphavantage API.

Modules

api

Module for basic definition of user information like setting API and requesting through that API

client

Module which provides trait to implement own client as well as default client in project

crypto

Module for crypto real time data

earning

Module for return earning per share for a company

error

Module which contains all types of error for alpha vantage crates

exchange

Module for exchange currency (both digital & physical currency exchange)

forex

Module for Forex realtime and historical data

quote

Module for returning latest price and volume information

search

Module for searching specific symbol or companies

sector

Module for sector

stock_time

Module for stock time series

technical_indicator

Module for Technical Indicator

Functions

set_api

Set API key using user selected or created client