Crate alpha_vantage

Rust Client/Wrapper built for Alphavantage API.

Modules

crypto
Module for crypto real time data

crypto_rating
Module for getting crypto health rating provided by FCAS metric

earning
Module for return earning per share for a company

error
Module which contains all types of error for alpha vantage crates

exchange
Module for exchange currency (both digital & physical currency exchange)

forex
Module for Forex realtime and historical data

quote
Module for returning latest price and volume information

search
Module for searching specific symbol or companies

sector
Module for sector

stock_time
Module for stock time series

technical_indicator
Module for Technical Indicator

user
Module for basic definition of user information like setting API and requesting through that API

util
Utility module declaring enum for basic function and parameters for different API

Functions

set_api
Set API value which can be used for calling different module