Struct alpha_vantage::api::ApiClient

pub struct ApiClient { /* private fields */ }

Struct for initializing client which contains different method for API call

Implementations

impl ApiClient

pub fn set_api<S, T>(api: S, client: T) -> Selfwhere
S: Into<String>,
T: HttpClient + 'static + Send + Sync,

Method for initializing ApiClient struct using user provided client and alphavantage.co provider

use alpha_vantage::api::ApiClient;
let api = ApiClient::set_api("some_key", reqwest::Client::new());

pub fn set_rapid_api<S, T>(api: S, client: T) -> Selfwhere
S: Into<String>,
T: HttpClient + 'static + Send + Sync,

Method for initializing ApiClient struct using user provided client and RapidAPI API provider

use alpha_vantage::api::ApiClient;
let api = ApiClient::set_api("some_key", reqwest::Client::new());

pub fn get_api_key(&self) -> &str

Method to get api key

use alpha_vantage::api::ApiClient;
let api = alpha_vantage::api::ApiClient::set_api("some_key", reqwest::Client::new());
assert_eq!(api.get_api_key(), "some_key");

pub fn crypto<'a>(
&'a self,
function: CryptoFunction,
symbol: &'a str,
market: &'a str
) -> CryptoBuilder<'a>

Crypto method for calling cryptography function with help of CryptoBuilder

Example

#[tokio::main]
async fn main() {
    let api = alpha_vantage::set_api("demo", reqwest::Client::new());
    let crypto = api
        .crypto(alpha_vantage::crypto::CryptoFunction::Daily, "BTC", "CNY")
        .json()
        .await
        .unwrap();
    assert_eq!(crypto.digital_code(), "BTC");
    assert_eq!(crypto.digital_name(), "Bitcoin");
    assert_eq!(crypto.market_code(), "CNY");
    assert_eq!(crypto.market_name(), "Chinese Yuan");
}

pub fn custom<'a>(&'a self, function: &'a str) -> CustomBuilder<'a>

Method for calling custom function not implemented currently in library using CustomBuilder

pub fn earning<'a>(&'a self, symbol: &'a str) -> EarningBuilder<'a>

Method for returning EarningBuilder for earning API

Example

#[tokio::main]
async fn main() {
    let api = alpha_vantage::set_api("demo", reqwest::Client::new());
    let earning = api.earning("IBM").json().await.unwrap();
    let symbol = earning.symbol();
    assert_eq!(symbol, "IBM");
}

pub fn economic_indicator<'a>(
&'a self,
function: &'a str
) -> EconomicIndicatorBuilder<'a>

Method for economic indicator builder

Example

#[tokio::main]
async fn main() {
    let api = alpha_vantage::set_api("demo", reqwest::Client::new());
    let economic = api
        .economic_indicator("REAL_GDP_PER_CAPITA")
        .json()
        .await
        .unwrap();
    assert_eq!(economic.interval(), "quarterly");
}

pub fn exchange<'a>(
&'a self,
from_currency: &'a str,
to_currency: &'a str
) -> ExchangeBuilder<'a>

Method for creating ExchangeBuilder for exchanging currency value from one currency to another currency.

Example

#[tokio::main]
async fn main() {
    let api = alpha_vantage::set_api("demo", reqwest::Client::new());
    let exchange = api.exchange("BTC", "CNY").json().await.unwrap();
    assert_eq!(exchange.name_from(), "Bitcoin");
    assert_eq!(exchange.code_from(), "BTC");
    assert_eq!(exchange.name_to(), "Chinese Yuan");
    assert_eq!(exchange.code_to(), "CNY");
}

pub fn forex<'a>(
&'a self,
function: ForexFunction,
from_symbol: &'a str,
to_symbol: &'a str
) -> ForexBuilder<'a>

Method for creating ForexBuilder for Forex API

Example

#[tokio::main]
async fn main() {
    let api = alpha_vantage::set_api("demo", reqwest::Client::new());
    let forex = api
        .forex(alpha_vantage::forex::ForexFunction::Weekly, "EUR", "USD")
        .json()
        .await
        .unwrap();
    assert_eq!(forex.symbol_from(), "EUR");
    assert_eq!(forex.symbol_to(), "USD");
    assert!(forex.interval().is_none());
}

pub fn quote<'a>(&'a self, symbol: &'a str) -> QuoteBuilder<'a>

Method for creating QuoteBuilder from APIClient

Example

#[tokio::main]
async fn main() {
    let api = alpha_vantage::set_api("demo", reqwest::Client::new());
    let quote = api.quote("MSFT").json().await.unwrap();
    let symbol = quote.symbol();
    assert_eq!(symbol, "MSFT");
}

pub fn search<'a>(&'a self, keywords: &'a str) -> SearchBuilder<'a>

Method for creating search builder

Example

#[tokio::main]
async fn main() {
    let api = alpha_vantage::set_api("demo", reqwest::Client::new());
    let search = api.search("BA").json().await.unwrap();
    let first_search_match = &search.matches()[0];
    assert_eq!(first_search_match.symbol(), "BA");
    assert_eq!(first_search_match.name(), "Boeing Company");
    assert_eq!(first_search_match.stock_type(), "Equity");
    assert_eq!(first_search_match.region(), "United States");
    assert_eq!(first_search_match.currency(), "USD");
    assert_eq!(first_search_match.match_score(), 1.0);
}

pub fn sector(&self) -> SectorBuilder<'_>

Method for creating SectorBuilder

Example

#[tokio::main]
async fn main() {
    let api = alpha_vantage::set_api("demo", reqwest::Client::new());
    let sector = api.sector().json().await.unwrap();
    assert_eq!(
        sector.information(),
        "US Sector Performance (realtime & historical)"
    );
}

pub fn stock_time<'a>(
&'a self,
function: StockFunction,
symbol: &'a str
) -> TimeSeriesBuilder<'a>

Method for creating Stock time Builder from APIClient

Example

#[tokio::main]
async fn main() {
    let api = alpha_vantage::set_api("demo", reqwest::Client::new());
    let stock = api
        .stock_time(alpha_vantage::stock_time::StockFunction::Weekly, "MSFT")
        .json()
        .await
        .unwrap();
    assert_eq!(stock.symbol(), "MSFT");
    assert!(stock.interval().is_none());
}

pub fn technical_indicator<'a>(
&'a self,
function: &'a str,
symbol: &'a str,
interval: TechnicalIndicatorInterval
) -> TechnicalIndicatorBuilder<'a>

Method for technical indicator builder

Example

#[tokio::main]
async fn main() {
    let api = alpha_vantage::set_api("demo", reqwest::Client::new());
    let technical = api
        .technical_indicator(
            "MAMA",
            "IBM",
            alpha_vantage::technical_indicator::TechnicalIndicatorInterval::Daily,
        )
        .series_type("close")
        .extra_param("fastlimit", 0.02)
        .json()
        .await;
    assert!(technical.is_ok());
}