Struct alpha_vantage::api::ApiClient

pub struct ApiClient { /* fields omitted */ }

Struct for initializing client which contains different method for API call

Implementations

impl ApiClient

#[must_use]
pub fn set_api<T>(api: &str, client: T) -> Self where
T: HttpClient + 'static,

Method for initializing ApiClient struct using user provided client

use alpha_vantage::api::ApiClient;
let api = ApiClient::set_api("some_key", surf::Client::new());

#[must_use]
pub fn get_api(&self) -> &str

Method to get api key

use alpha_vantage::api::ApiClient;
let api = alpha_vantage::api::ApiClient::set_api("some_key", surf::Client::new());
assert_eq!(api.get_api(), "some_key");

pub async fn crypto_rating(&self, symbol: &str) -> Result<CryptoRating>

Method for getting crypto health rating

Example

#[async_std::main]
async fn main() {
    let api = alpha_vantage::set_api("demo", surf::Client::new());
    let crypto_rating = api.crypto_rating("BTC").await.unwrap();
    assert_eq!(crypto_rating.symbol(), "BTC");
    assert_eq!(crypto_rating.name(), "Bitcoin");
    assert_eq!(crypto_rating.time_zone(), "UTC");
}

pub async fn crypto(
&self,
function: CryptoFunction,
symbol: &str,
market: &str
) -> Result<Crypto>

Crypto method for calling cryptography function

Example

#[async_std::main]
async fn main() {
    let api = alpha_vantage::set_api("demo", surf::Client::new());
    let crypto = api
        .crypto(alpha_vantage::utils::CryptoFunction::Daily, "BTC", "CNY")
        .await
        .unwrap();
    assert_eq!(crypto.digital_code(), "BTC");
    assert_eq!(crypto.digital_name(), "Bitcoin");
    assert_eq!(crypto.market_code(), "CNY");
    assert_eq!(crypto.market_name(), "Chinese Yuan");
}

pub async fn custom<T>(
&self,
function: &str,
extras: Vec<(&str, &str)>
) -> Result<T> where
T: DeserializeOwned,

Method for calling custom function not implemented currently in library

pub async fn earning(&self, symbol: &str) -> Result<Earning>

Method for returning company earning

Example

#[async_std::main]
async fn main() {
    let api = alpha_vantage::set_api("demo", surf::Client::new());
    let earning = api.earning("IBM").await.unwrap();
    let symbol = earning.symbol();
    assert_eq!(symbol, "IBM");
}

pub async fn exchange(
&self,
from_currency: &str,
to_currency: &str
) -> Result<Exchange>

Method for exchanging currency value from one currency to another currency.

Example

#[async_std::main]
async fn main() {
    let api = alpha_vantage::set_api("demo", surf::Client::new());
    let exchange = api.exchange("BTC", "CNY").await.unwrap();
    assert_eq!(exchange.name_from(), "Bitcoin");
    assert_eq!(exchange.code_from(), "BTC");
    assert_eq!(exchange.name_to(), "Chinese Yuan");
    assert_eq!(exchange.code_to(), "CNY");
}

pub async fn forex(
&self,
function: ForexFunction,
from_symbol: &str,
to_symbol: &str,
interval: TimeSeriesInterval,
output_size: OutputSize
) -> Result<Forex>

Method for calling stock time series forex Api

Example

use alpha_vantage::utils::*;
#[async_std::main]
async fn main() {
    let api = alpha_vantage::set_api("demo", surf::Client::new());
    let forex = api
        .forex(
            ForexFunction::Weekly,
            "EUR",
            "USD",
            TimeSeriesInterval::None,
            OutputSize::None,
        )
        .await
        .unwrap();
    assert_eq!(forex.symbol_from(), "EUR");
    assert_eq!(forex.symbol_to(), "USD");
    assert!(forex.interval().is_none());
}

pub async fn quote(&self, symbol: &str) -> Result<Quote>

Method for returning Quote Struct

Example

#[async_std::main]
async fn main() {
    let api = alpha_vantage::set_api("demo", surf::Client::new());
    let quote = api.quote("MSFT").await.unwrap();
    let symbol = quote.symbol();
    assert_eq!(symbol, "MSFT");
}

pub async fn search(&self, keywords: &str) -> Result<Search>

Method for searching keyword or company

Example

#[async_std::main]
async fn main() {
    let api = alpha_vantage::set_api("demo", surf::Client::new());
    let search = api.search("BA").await.unwrap();
    let first_search_result = &search.result()[0];
    assert_eq!(first_search_result.symbol(), "BA");
    assert_eq!(first_search_result.name(), "Boeing Company");
    assert_eq!(first_search_result.stock_type(), "Equity");
    assert_eq!(first_search_result.region(), "United States");
    assert_eq!(first_search_result.currency(), "USD");
    assert_eq!(first_search_result.match_score(), 1.0);
}

pub async fn sector(&self) -> Result<Sector>

Method for returning a sector data as struct

Example

#[async_std::main]
async fn main() {
    let api = alpha_vantage::set_api("demo", surf::Client::new());
    let sector = api.sector().await.unwrap();
    assert_eq!(
        sector.information(),
        "US Sector Performance (realtime & historical)"
    );
}

pub async fn stock_time(
&self,
function: StockFunction,
symbol: &str,
interval: TimeSeriesInterval,
output_size: OutputSize
) -> Result<TimeSeries>

Method for calling stock time series API

Example

use alpha_vantage::utils::*;
#[async_std::main]
async fn main() {
    let api = alpha_vantage::set_api("demo", surf::Client::new());
    let stock = api
        .stock_time(
            StockFunction::Weekly,
            "MSFT",
            TimeSeriesInterval::None,
            OutputSize::None,
        )
        .await
        .unwrap();
    assert_eq!(stock.symbol(), "MSFT");
    assert!(stock.interval().is_none());
}

pub async fn technical_indicator(
&self,
function: &str,
symbol: &str,
interval: TechnicalIndicatorInterval,
time_period: Option<u64>,
series_type: Option<&str>,
extras: Vec<TechnicalIndicator>
) -> Result<Indicator>

Method for technical indicator API

Example

#[async_std::main]
async fn main() {
    let api = alpha_vantage::set_api("demo", surf::Client::new());
    let technical = api
        .technical_indicator(
            "SMA",
            "IBM",
            alpha_vantage::utils::TechnicalIndicatorInterval::Weekly,
            Some(10),
            Some("open"),
            vec![],
        )
        .await;
    assert!(technical.is_ok());
}