Trait alpha_vantage::client::HttpClient

pub trait HttpClient {
    fn get_output<'life0, 'async_trait>(
        &'life0 self, 
        path: String
    ) -> Pin<Box<dyn Future<Output = Result<String>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

Trait which can be implemented for all common library client for getting output from server surf and reqwest are two client which are supported with feature flag. If you prefer alternate http client you can add support by implementing HttpClient trait for client. Some example of other client which can be used are isahc client

Required methods

fn get_output<'life0, 'async_trait>(
&'life0 self,
path: String
) -> Pin<Box<dyn Future<Output = Result<String>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Get output from server in String. This string is automatically converted to appropriate struct by library