Trait alpha_vantage::crypto::VecEntry

pub trait VecEntry {
    fn find(&self, time: &str) -> Option<&Entry>;
    fn latest(&self) -> Entry;
    fn latest_n(&self, n: usize) -> Result<Vec<&Entry>>;
}

trait which helps for performing some common operation on Vec

Required Methods

fn find(&self, time: &str) -> Option<&Entry>

Find a entry with a given time as a input return none if no entry found

fn latest(&self) -> Entry

Return a entry which is of latest time period

fn latest_n(&self, n: usize) -> Result<Vec<&Entry>>

Return a top n latest Entry

Errors

If n is greater than no of entry