Module alpha_vantage::forex

Module for Forex realtime and historical data

APIs under this section provide a wide range of data feed for realtime and historical forex (FX) rates.

You can read about Forex API and what it returns on alphavantage documentation

Structs

Entry

Struct to store Entry value

Forex

Struct to store Forex data after forex API call

ForexBuilder

Builder to create Forex

Enums

ForexFunction

Enum for declaring function for forex function by defining which type of forex series to be returned

Traits

VecEntry

trait which helps for performing some common operation on Vec