Trait alpha_vantage::vec_trait::FindData

pub trait FindData: IntoIterator {
    // Required methods
    fn find(&self, time: &str) -> Option<&<Self as IntoIterator>::Item>;
    fn latest(&self) -> <Self as IntoIterator>::Item;
    fn latest_n(&self, n: usize) -> Result<Vec<&<Self as IntoIterator>::Item>>;
}

trait which helps for performing some common operation on Vec<Data> which have given period

Required Methods

fn find(&self, time: &str) -> Option<&<Self as IntoIterator>::Item>

Find a data with a given time as a input return none if no data found

fn latest(&self) -> <Self as IntoIterator>::Item

Return a data which is of latest time period

fn latest_n(&self, n: usize) -> Result<Vec<&<Self as IntoIterator>::Item>>

Return a top n latest Data

Errors

If n is greater than no of data